requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
ecdsa==0.18.0
fastapi==0.89.1
h11==0.14.0
idna==3.4
motor==3.1.1
passlib==1.7.4
pyasn1==0.4.8
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
